requirements.txt
requests
numpy
pandas
xgboost
scikit-learn
pynacl
